requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.92.0
anyio==4.13.0
attrs==26.1.0
blockbuster==1.5.26
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpickle==3.1.2
croniter==6.2.2
cryptography==46.0.7
dataclasses-json==0.6.7
distro==1.9.0
docstring_parser==0.17.0
forbiddenfruit==0.1.4
frozenlist==1.8.0
googleapis-common-protos==1.74.0
grpcio==1.78.0
grpcio-health-checking==1.78.0
grpcio-tools==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema_rs==0.44.1
langchain==1.2.15
langchain-anthropic==1.4.0
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.28
langchain-tavily==0.2.17
langchain-text-splitters==1.1.1
langgraph==1.1.6
langgraph-api==0.7.98
langgraph-checkpoint==4.0.1
langgraph-cli==0.4.21
langgraph-prebuilt==1.0.9
langgraph-runtime-inmem==0.27.3
langgraph-sdk==0.3.13
langsmith==0.7.27
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.4
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pathspec==1.0.4
propcache==0.4.1
protobuf==6.33.6
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
setuptools==82.0.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.49
sse-starlette==3.3.4
starlette==1.0.0
structlog==25.5.0
tavily-python==0.7.23
tenacity==9.1.4
tiktoken==0.12.0
truststore==0.10.4
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.44.0
watchfiles==1.1.1
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
